renaming epetra-default-setter
